+++ src/commands/youtube/index.ts
@@ -1,0 +1,3 @@
+import { YouTubeCommand } from './youtube';
+
+export default YouTubeCommand;